refactor(src, proto): Рефактор и сортировка функционала
