+++ Resources/Prototypes/ADT/Actions/phantom.yml
@@ -238,7 +238,7 @@
       sprite: ADT/Interface/Actions/actions_phantom.rsi
       state: oath
     itemIconStyle: BigAction
-    event: !type:PuppeterActionEvent
+    event: !type:PhantomOathActionEvent
     useDelay: 240
     priority: -78
